♻️ refactor: Provider를 GlobalProvider로 변경 및 적용 [#24]
