+++ src/app/provider/globalProvider.tsx
@@ -1,21 +1,25 @@
+import React from 'react';
+
 import { Global } from '@emotion/react';
 import { globalStyle } from '@/shared/config/styles/global';
 import { QueryClientProvider } from '@/app/provider/queryClientProvider';
 
 import ModalLayer from '@/shared/ui/modalLayer';
 
-import Home from '@/pages/home/home.ui';
+interface IGlobalProviderProps {
+  children: React.ReactNode;
+}
 
-const Provider = () => {
+export const GlobalProvider: React.FC<IGlobalProviderProps> = ({
+  children,
+}) => {
   return (
     <QueryClientProvider>
       <ModalLayer.Provider>
         <Global styles={globalStyle} />
-        <Home />
+        {children}
         <ModalLayer.Wrapper />
       </ModalLayer.Provider>
     </QueryClientProvider>
   );
 };
-
-export default Provider;
